[#143]refact: MainCardItem 컴포넌트 분리
