+++ src/components/layout/MainCardList.tsx
@@ -35,10 +35,10 @@ const MainCardList = ({ tab }: IMainCardListProps) => {
         page.map(card => (
           <MainCardItem
             key={card.id}
-            id={String(card.id)}
+            id={card.id}
             name={card.name}
-            dateTime={new Date(card.dateTime)}
-            registrationEnd={new Date(card.registrationEnd)}
+            dateTime={card.dateTime}
+            registrationEnd={card.registrationEnd}
             location={card.location}
             participantCount={card.participantCount}
             capacity={card.capacity}
